refactor: centralize CLI child execution labels
