+++ src/test-kernel/cli/ChildControls.vitest.mts
@@ -36,7 +36,7 @@ import {
   resolveChildControlDisplayTargets,
   resolveChildControlStreamTarget,
 } from '@cli/chat/tui/state/childControls';
-import { visibleSubagentRows } from '@cli/chat/tui/state/childStreamMerge';
+import { visibleSubagentRows } from '@cli/chat/tui/state/childExecutions';
 import { NO_BYPASS } from '@cli/chat/tui/state/cliState';
 import { streamDisplayLabel } from '@cli/chat/tui/state/streamViews';
 import type {
